requirements.txt
flask
requests
pyyaml
jinja2